+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorArrayTest.kt
@@ -65,8 +65,11 @@ class CodeGeneratorArrayTest {
 
 import java.util.UUID
 
+/**
+ * Test generation of arrays.
+ */
 data class TestArray(
-        val aaa: List<Person>
+    val aaa: List<Person>
 ) {
 
     init {
@@ -75,8 +78,8 @@ data class TestArray(
     }
 
     data class Person(
-            val id: UUID,
-            val name: String
+        val id: UUID,
+        val name: String
     ) {
 
         init {
@@ -99,6 +102,9 @@ import java.util.List;
 import java.util.UUID;
 import java.util.regex.Pattern;
 
+/**
+ * Test generation of arrays.
+ */
 public class TestArray {
 
     private static final Pattern cg_regex0 = Pattern.compile("^[A-Z][A-Za-z]*${'$'}");