+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorCustomExtraTest.kt
@@ -92,13 +92,13 @@ class CodeGeneratorCustomExtraTest {
 import java.math.BigDecimal
 
 data class Test(
-        /** Product identifier */
-        val id: BigDecimal,
-        /** Name of the product */
-        val name: String,
-        val price: BigDecimal,
-        val tags: List<String>? = null,
-        val stock: Stock? = null
+    /** Product identifier */
+    val id: BigDecimal,
+    /** Name of the product */
+    val name: String,
+    val price: BigDecimal,
+    val tags: List<String>? = null,
+    val stock: Stock? = null
 ) {
 
     init {
@@ -118,8 +118,8 @@ data class Test(
 import java.math.BigDecimal
 
 data class Stock(
-        val warehouse: BigDecimal? = null,
-        val retail: BigDecimal? = null
+    val warehouse: BigDecimal? = null,
+    val retail: BigDecimal? = null
 )
 """