+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorDefaultTest.kt
@@ -53,11 +53,14 @@ class CodeGeneratorDefaultTest {
 """package com.example
 
 
+/**
+ * Test default values.
+ */
 data class TestDefault(
-        val aaa: Long = 8,
-        val bbb: String? = null,
-        val ccc: String = "CCC",
-        val ddd: List<Long> = listOf(123, 456)
+    val aaa: Long = 8,
+    val bbb: String? = null,
+    val ccc: String = "CCC",
+    val ddd: List<Long> = listOf(123, 456)
 )
 """