+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorEmptyClassTest.kt
@@ -63,12 +63,18 @@ class CodeGeneratorEmptyClassTest {
         const val expected =
 """package com.example
 
+/**
+ * Test empty object.
+ */
 class TestEmpty
 """
 
         const val expectedJava =
 """package com.example;
 
+/**
+ * Test empty object.
+ */
 public class TestEmpty {
 
     @Override