+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorExampleTest.kt
@@ -66,13 +66,13 @@ class CodeGeneratorExampleTest {
 import java.math.BigDecimal
 
 data class Test(
-        /** Product identifier */
-        val id: BigDecimal,
-        /** Name of the product */
-        val name: String,
-        val price: BigDecimal,
-        val tags: List<String>? = null,
-        val stock: Stock? = null
+    /** Product identifier */
+    val id: BigDecimal,
+    /** Name of the product */
+    val name: String,
+    val price: BigDecimal,
+    val tags: List<String>? = null,
+    val stock: Stock? = null
 ) {
 
     init {
@@ -80,8 +80,8 @@ data class Test(
     }
 
     data class Stock(
-            val warehouse: BigDecimal? = null,
-            val retail: BigDecimal? = null
+        val warehouse: BigDecimal? = null,
+        val retail: BigDecimal? = null
     )
 
     companion object {